tests/wpt/web-platform-tests/css/css-ui/parsing/box-sizing-computed.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 3: getComputedStyle().boxSizing</title>
<link rel="help" href="https://drafts.csswg.org/css-ui-3/#box-sizing">
<meta name="assert" content="box-sizing computed value is as specified.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/computed-testcommon.js"></script>
</head>
<body>
<div id="target"></div>
<script>
test_computed_value("box-sizing", "content-box");
test_computed_value("box-sizing", "border-box");
</script>
</body>
</html>
